Implement first iteration of chord analysis algorithm
tldr we assume that the root of the chord is present.
From there we brute force every single possible chord (out of the
supported chords) and pick the one with the most matching notes. In
the event of a tie we look at the lowest note and check if that's one
of the options for chord roots.